fix typos

+++ controllers/api/comment-routes.js
@@ -1,24 +1,12 @@
 const router = require('express').Router();
-const { Comment, User } = require('../../models');
+const { Comment } = require('../../models');
 
 router.get('/', (req, res) => {
-    Comment.findAll({
-        attributes: [
-            'id',
-            'comment_text',
-            'user_id',
-            'post_id',
-            'created_at'
-        ],
-        include: {
-            model: User,
-            attributes: ['username']
-        }
-    })
+    Comment.findAll()
     .then(dbCommentData => res.json(dbCommentData))
     .catch(err => {
         console.log(err);
-        res.status(400).json(err);
+        res.status(500).json(err);
     });
 });
 
@@ -40,21 +28,24 @@ router.post('/', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
-    Comment.destroy({
+    if (req.session) {
+      Comment.destroy({
         where: {
-            id: req.params.id
+          id: req.params.id
         }
-    })
-    .then(dbCommentData => {
-        if (!dbCommentData) {
-            res.status(404).json({ message: 'No comment found with this id'});
-        }
-        res.json(dbCommentData);
-    })
-    .catch(err => {
-        console.log(err);
-        res.status(500).json(err);
-    })
+      })
+        .then(dbCommentData => {
+            if (!dbCommentData) {
+            res.status(404).json({ message: 'No comment found with this id!' });
+            return;
+            }
+            res.json(dbCommentData);
+        })
+        .catch(err => {
+            console.log(err);
+            res.status(500).json(err);
+        });
+    }
 });
 
 module.exports = router;
